fix(mutual-credit): address CodeRabbit review on PR #2176
Validation & encoding
- validation.ts: tighten defaultCreditLimit to z.number().int() so fractional
  inputs are rejected at form time rather than throwing inside integerBigInt
  during tx encoding.
- useIssueNewTokenOrchestrator: throw when web3SpaceId is missing/non-positive
  before adding it to creditWhitelistSpaceIds so we never encode a 0/falsy id
  via batchAddCreditWhitelistSpaces.
- issue-new-token-form: gate the credit-limit superRefine on isRegularToken to
  match the UI gating, and refuse to submit a credit-enabled token if
  web3SpaceId is invalid.

On-chain reads (precision/safety)
- useTokenOnChainData: use formatUnits for maxSupply/defaultCreditLimit so
  sub-token values aren't truncated; warn when a uint256 space id exceeds
  Number.MAX_SAFE_INTEGER.
- get-mutual-credit-info: log partial multicall failures (not just
  all-failed); only fall back to creditLimit-creditBalance arithmetic when
  both inputs actually succeeded so partial failures can't overstate
  available credit.

Proposal cards
- proposal-token-items / proposal-update-token: use formatUnits to display
  defaultCreditLimit so 0.5-token limits no longer collapse to 0; localize
  the space-logo alt text via ProposalDetails.labels.spaceLogoAlt.

Asset card / list
- asset-card: skip the net-balance Badge when netBalance === 0, and avoid
  trailing-space rendering by gating the whole block on a defined symbol.
- assets-list: re-export the canonical AssetItem from use-user-assets and
  drop the divergent local definition.

Transfer form
- people-transfer-funds: only surface zero-balance credit-eligible tokens
  when creditLimitLeft > 0 so the dropdown can't suggest a transfer the
  submit-time check would immediately reject.

Update-token plugin
- clearAdvancedSettingsFields now also resets enableMutualCredit,
  defaultCreditLimit and creditWhitelistedSpaceIds so collapsing Advanced
  Settings can't smuggle hidden credit values into the update tx.

Mutual-credit section
- type useFormContext with a MutualCreditFormValues generic for compile-time
  field-name checking and drop the now-redundant `as number[]` casts.
- give the enable Switch an aria-labelledby pointing at the visible label
  for screen-reader announcement.

API route
- /people/[slug]/assets: replace `[] as any[]` with the proper
  Awaited<ReturnType<typeof getTokenBalancesByAddress>> fallback.

i18n
- en.json: add ProposalDetails.labels.spaceLogoAlt.
- de/es/fr/pt.json: add the missing TreasuryTab.assetCard.mutualCredit
  badges, AgreementFlow.issueNewTokenForm.errors.creditLimitPositive,
  AgreementFlow.plugins.issueNewToken.mutualCredit.* and the new
  ProposalDetails.labels mutual-credit/spaceLogoAlt keys.

Type-narrowing fallout from the AssetItem dedupe (address became optional)
- guard asset.address in people-redeem-tokens, people-transfer-form,
  pending-rewards-section and use-user-assets-section.

Made-with: Cursor

+++ packages/core/src/governance/validation.ts
@@ -655,7 +655,7 @@ export const baseSchemaIssueNewToken = z.object({
     }
     const num = Number(val);
     return isNaN(num) ? undefined : num;
-  }, z.number().nonnegative().optional()),
+  }, z.number().int().nonnegative().optional()),
   /** Web3 space ids whose members are eligible for the credit line */
   creditWhitelistedSpaceIds: z.array(z.number().int().nonnegative()).optional(),
 });
